Rebind more keys in Dyson Sphere Program
This mod makes it possible for me to rebind some hard-coded keys in the
game, such as `E` opening inventory, or closing panels, that otherwise
make the game difficult to play with a dvorak layout.

This is an incomplete version: it allows rebinding of `E`, but does not
fully address the behaviour since the runtime UI code has hard-coded
the binding.

Since this is currently failing at runtime due to the
`System.Reflection.Emit.ILGeneration` Assembly being missing, this is
being published to get feedback from more experienced DSP modders.